+++ addOns/exim/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Add Automation Framework job to export data (e.g. HAR, URLs).
+- Support for Sites Tree export and prune.
 
 ### Changed
 - Update dependency.